test(api): added tests to User.test.ts
